Tema 3/hello_world/requirements.txt
Flask==1.1.2
future==0.18.2
gunicorn==20.1.0
google-cloud-storage==1.37.0
flask-cors==3.0.10
six==1.15.0
pyyaml==5.4.1
requests==2.25.1
google-auth==1.28.0
google-auth-oauthlib==0.4.4
google-cloud-translate==2.0.1
google-cloud-texttospeech==2.3.0
